refactoring into roomservice
